gitignore and indent change

+++ contracts/Game.sol
@@ -14,35 +14,34 @@ contract GameFactory {
 }
 
 contract Game {
-
-    address manager;    
-    uint minimum_amount;
+    address manager;
+    uint256 minimum_amount;
     address[] public players;
-    mapping(address => uint) public players_map;
-    
+    mapping(address => uint256) public players_map;
+
     modifier restricted() {
         require(players_map[msg.sender] >= minimum_amount);
         _;
     }
-    
-    function Game(uint minimum, address creator) public {
+
+    function Game(uint256 minimum, address creator) public {
         minimum_amount = minimum;
         manager = creator;
     }
-    
+
     function enter() public payable {
         require(msg.value >= minimum_amount);
-        
+
         players.push(msg.sender);
         players_map[msg.sender] = msg.value;
     }
-    
-    function get_players() public view returns(address[]) {
+
+    function get_players() public view returns (address[]) {
         return players;
     }
-    
+
     function end_game() public restricted {
-        for (uint i=0; i<players.length; i++) {
+        for (uint256 i = 0; i < players.length; i++) {
             address player = players[i];
             player.transfer(players_map[player]);
             players_map[player] = 0;
